+++ src/main/java/edu/urgu/oopteam/viewmodels/ExchangeData.java
@@ -1,22 +1,24 @@
 package edu.urgu.oopteam.viewmodels;
 
+import java.math.BigDecimal;
+
 /**
  * Holds currency exchange rate and associated bank name
  */
 public class ExchangeData {
     private String bankName;
-    private double rate;
+    private BigDecimal rate;
 
-    public ExchangeData(String bankName, double rate) {
+    public ExchangeData(String bankName, BigDecimal rate) {
         this.bankName = bankName;
         this.rate = rate;
     }
 
-    public double getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
 
-    public void setRate(double rate) {
+    public void setRate(BigDecimal rate) {
         this.rate = rate;
     }
